Add a bookmark

You must give a url and description, all other arguments are optional.

pb_posts_add(
  url,
  title,
  description = NULL,
  tags = NULL,
  dt = NULL,
  replace = NULL,
  public = NULL,
  toread = NULL,
  username = NULL,
  token = NULL
)

Arguments

url
REQUIRED the URL of the item

title
REQUIRED Title of the item.

description
Description of the item

tags
vector of up to 100 tags

dt
creation time for this bookmark. Defaults to current time. Datestamps more than 10 minutes ahead of server time will be reset to current server time (UTC timestamp in this format: 2010-12-11T19:48:02Z)

replace
Replace any existing bookmark with this URL. Default is TRUE If set to FALSE, will throw an error if bookmark exists

public
TRUE/FALSE Make bookmark public. Default is "TRUE" unless user has enabled the "save all bookmarks as private" user setting, in which case default is "no"

toread
TRUE/FALSE Marks the bookmark as unread. Default is "FALSE"

username
your username You probably don't need to supply this every time. See ?authentication

token
your token

Value

status in text 'done'

Examples

if (FALSE) {
pb_posts_add()
}